refactor(tooling): fold conversation-goldens miner review nits
Four non-blocking review items on the miner, done now rather than left to rot:

- extract the shared bucket-sampling helper (sampling.ts) so mine-goldens and
  mine-conversation-goldens can't drift the same even-spacing algorithm
  (reuse-scout consolidation; stratifySample now calls it, determinism held)
- fetchPriorHistory: (created_at, id) row-value tie-break so a same-ms prior
  message the channel history would include isn't silently dropped from the fold
- clarify the historyWindow comment — it's the raw over-fetch POOL the fold
  slices its tail from, not the fold size (which is LTM_SEARCH_HISTORY_TURNS=3)
- validate --persona-id as a UUID before the ::uuid cast (clean CLI error, not a
  raw Postgres cast error); shared requirePersonaId de-dupes the required check
